requirements.txt
Flask
tensorflow
opencv-python
numpy
Pillow
sqlalchemy
wtforms
flask-wtf
gunicorn
huggingface_hub
